вторая лаба готова

+++ src/main/java/ru/bmstu/rsoi/web/AuthorController.java
@@ -8,7 +8,12 @@ import ru.bmstu.rsoi.dto.PersonUpdateRequest;
 import ru.bmstu.rsoi.entity.Author;
 import ru.bmstu.rsoi.service.LibraryPersonService;
 
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+import java.io.IOException;
 import java.util.List;
+
+import static ru.bmstu.rsoi.web.OAuthChecker.checkOAuth;
 
 /**
  * Created by ali on 23.11.16.
@@ -24,12 +29,18 @@ public class AuthorController {
     private LibraryPersonService<Author> authorService;
 
     @RequestMapping(value = "/add", method = RequestMethod.PUT)
-    public @ResponseBody Author addAuthor(@RequestBody PersonUpdateRequest request) {
+    public @ResponseBody Author addAuthor(@RequestBody PersonUpdateRequest request,
+                                          HttpServletRequest httpRequest, HttpServletResponse response) throws IOException {
+        if (!checkOAuth(httpRequest, response))
+            return null;
         return authorService.mergePerson(null, request.getName(), request.getBornDate(), null);
     }
 
     @RequestMapping(value = "/{id}/remove", method = RequestMethod.DELETE)
-    public void removeAuthor(@PathVariable int id) {
+    public void removeAuthor(@PathVariable int id,
+                             HttpServletRequest httpRequest, HttpServletResponse response) throws IOException {
+        if (!checkOAuth(httpRequest, response))
+            return;
         authorService.removePerson(id);
     }
 
@@ -46,7 +57,10 @@ public class AuthorController {
     }
 
     @RequestMapping(value = "/{id}/update", method = RequestMethod.POST)
-    public @ResponseBody Author updateAuthor(@PathVariable int id, @RequestBody PersonUpdateRequest request) {
+    public @ResponseBody Author updateAuthor(@PathVariable int id, @RequestBody PersonUpdateRequest request,
+                                             HttpServletRequest httpRequest, HttpServletResponse response) throws IOException {
+        if (!checkOAuth(httpRequest, response))
+            return null;
         return authorService.mergePerson(id, request.getName(), request.getBornDate(), request.getVersion());
     }
 }
