+++ src/main/java/ru/bmstu/rsoi/web/VisitorController.java
@@ -8,7 +8,12 @@ import ru.bmstu.rsoi.dto.PersonUpdateRequest;
 import ru.bmstu.rsoi.entity.LibraryVisitor;
 import ru.bmstu.rsoi.service.LibraryPersonService;
 
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+import java.io.IOException;
 import java.util.List;
+
+import static ru.bmstu.rsoi.web.OAuthChecker.checkOAuth;
 
 /**
  * Created by ali on 23.11.16.
@@ -26,12 +31,18 @@ public class VisitorController {
     private LibraryPersonService<LibraryVisitor> visitorService;
 
     @RequestMapping(value = "add", method = RequestMethod.PUT)
-    public @ResponseBody LibraryVisitor addVisitor(@RequestBody PersonUpdateRequest request) {
+    public @ResponseBody LibraryVisitor addVisitor(@RequestBody PersonUpdateRequest request,
+                                                   HttpServletRequest httpRequest, HttpServletResponse response) throws IOException {
+        if (!checkOAuth(httpRequest, response))
+            return null;
         return visitorService.mergePerson(null, request.getName(), request.getBornDate(), null);
     }
 
     @RequestMapping(value = "/{id}/remove", method = RequestMethod.DELETE)
-    public void removeVisitor(@PathVariable int id) {
+    public void removeVisitor(@PathVariable int id,
+                              HttpServletRequest httpRequest, HttpServletResponse response) throws IOException {
+        if (!checkOAuth(httpRequest, response))
+            return;
         visitorService.removePerson(id);
     }
 
@@ -48,7 +59,10 @@ public class VisitorController {
     }
 
     @RequestMapping(value = "/{id}/update", method = RequestMethod.POST)
-    public LibraryVisitor updateVisitor(@PathVariable int id, @RequestBody PersonUpdateRequest request) {
+    public LibraryVisitor updateVisitor(@PathVariable int id, @RequestBody PersonUpdateRequest request,
+                                        HttpServletRequest httpRequest, HttpServletResponse response) throws IOException {
+        if (!checkOAuth(httpRequest, response))
+            return null;
         return visitorService.mergePerson(id, request.getName(), request.getBornDate(), request.getVersion());
     }
 }
